automated_resume_shortlisting/resume_processor.py
import PyPDF2
import docx

def process_resume(file):
    # Extract text based on the file type
    if file.type == "application/pdf":
        return extract_text_from_pdf(file)
    elif file.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
        return extract_text_from_docx(file)

def extract_text_from_pdf(pdf_file):
    # Read and extract text from the PDF
    reader = PyPDF2.PdfReader(pdf_file)
    text = ""
    for page in reader.pages:
        text += page.extract_text()
    return text

def extract_text_from_docx(docx_file):
    # Read and extract text from DOCX
    doc = docx.Document(docx_file)
    text = ""
    for para in doc.paragraphs:
        text += para.text + "\n"
    return text
