automated_resume_shortlisting/requirements.txt
streamlit
pandas
scikit-learn
docx2txt
PyMuPDF
numpy